// src/interfaces.ts
export interface IPInfo {
	ip: string;
	city: string;
	region: string;
	region_code: string;
	country: string;
	country_code: string;
	region_plus_code: string;
  }

export interface IPResponse {
	ip: string,
	city: string,
	region: string,
	region_code: string,
	country: string,
	country_code: string,
  }